Renames Contentful\Delivery\Link to Contentful\Link.
This will aid a future CMA SDK.
